# openapi.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  description: API cho hệ thống đặt hàng đa NCC (Multi-Supplier Order). Chi tiết xem API-FE-LIEN-KET.md
  version: 1.0.0
servers:
  - url: /api
    description: Base path API
...
paths:
...
  /suppliers:
    get:
      tags: [Suppliers]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [Active, Inactive] }
      security: [{ Bearer: [] }]
      responses:
        '200':
          description: Danh sách NCC
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Supplier' }
    post:
      tags: [Suppliers]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplierInput' }
      security: [{ Bearer: [] }]
      responses:
        '201':
          description: Đã tạo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Supplier' }
  /suppliers/{id}:
    put:
      tags: [Suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SupplierInput' }
      security: [{ Bearer: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Supplier' }
    delete:
      tags: [Suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '204': { description: Đã xóa }
...
  /users:
    get:
      tags: [Users]
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
    post:
      tags: [Users]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserInput' }
      security: [{ Bearer: [] }]
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /users/{id}:
    put:
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserInput' }
      security: [{ Bearer: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    delete:
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      security: [{ Bearer: [] }]
      responses:
        '204': { description: Đã xóa }
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token trả về từ POST /auth/login
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [Admin, StoreUser, SupplierUser] }
        storeId: { type: integer, nullable: true }
        supplierId: { type: integer, nullable: true }
        status: { type: string }
    UserInput:
      type: object
      properties:
        email: { type: string }
        name: { type: string }
        role: { type: string }
        storeId: { type: integer, nullable: true }
        supplierId: { type: integer, nullable: true }
        status: { type: string }
    Supplier:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        contact: { type: string }
        email: { type: string }
        address: { type: string }
        status: { type: string }
    SupplierInput:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        contact: { type: string }
        email: { type: string }
        address: { type: string }
        status: { type: string }
...